refactor(ipc): extract dictation command module
